requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==23.1.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
cssselect==1.2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.6.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
parsel==1.8.1
pefile==2023.2.7
Pillow==9.5.0
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pyOpenSSL==23.1.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PySocks==1.7.1
pywin32==306
pywin32-ctypes==0.2.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
Scrapy==2.7.1
scrapy-rotating-proxies==0.6.2
selenium==4.9.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tldextract==3.4.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing==3.7.4.3
typing_extensions==4.5.0
urllib3==1.26.15
w3lib==2.1.1
wsproto==1.2.0
zipp==3.15.0
zope.interface==6.0
